Add possibility to drag tasks in columns to sort them
Implemented feature that task cards can be dragged within a column, and make the resulting order stick. This was done by setting the Ordinal property of the TaskItem.

Added a new config parameter to enable or disable this feature, "USE_ORDINALS". If this is true, sorting on ordinals is forced and the other sorting config of individual columns will be ignored (it doesn't make sense to set Ordinal and then sort on something else).

Fixed bug with INCLUDE_TODOS where dragging a task would attempt to move it to another folder, which failed. Don't attempt to move tasks at all when INCLUDE_TODOS is set since all tasks come from the same "folder" in this case and it is not possible to have different subfolders when using this feature anyway.

Changed default config slightly to avoid e-mails flagged as completed to appear in columns when INCLUDE_TODOS is enabled.

Some small comment fixes.