Fix Add Task buttons when using INCLUDE_TODOS
It now works to add new tasks to the board when using INCLUDE_TODOS. It
didn't work before because that folder is not allowed to Add to.

Sometimes the new task does not appear without a manual reload. Perhaps
it is a race condition between the reload on event AfterWrite and the
addition of the task to the To-Do List. Not sure. I have tried using
other events like Items.ItemAdd but have not managed to make it work.