Add support to drag mail items between columns
It is now possible to drag flagged mail items between the columns and
they will stick to the dropped column. This is accomplished by setting
the FlagRequest property. It is possible to drag it to and from the
Completed column as well with expected results.

Renamed config parameter INCLUDE_TODOS_STATUS to INITIAL_MAIL_STATUS.